+++ templates/listar.html
@@ -15,6 +15,10 @@
                 <td>{{ musica.titulo }}</td>
                 <td>{{ musica.categoria }}</td>
                 <td>{{ musica.idioma }}</td>
+                <td>
+                    <a href="{{ url_for('editar', id=musica.id) }}">Editar</a>
+                    <a href="{{ url_for('eliminar', id=musica.id) }}">Eliminar</a>
+                </td>
             </tr>
             {% endfor %}
         </tbody>
